% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_criterion.R
\name{inf_criterion}
\alias{inf_criterion}
\title{Information Criterion}
\usage{
inf_criterion(model, x, y, family = "binomial", lambda = "lambda.1se",
  metric = "EBIC")
}
\arguments{
\item{model}{Regression model.}

\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{family}{GLM family.}

\item{lambda}{Used in cv.glmnet.}

\item{metric}{AIC,  AICc, BIC or EBIC.}
}
\description{
Information Criterion
}
