% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid_adj.R
\name{vivid_adj}
\alias{vivid_adj}
\title{Adjust variables in a VIVID object}
\usage{
vivid_adj(vividObj, minFinalFeatures)
}
\arguments{
\item{vividObj}{An object passed from the vivid() funciton.}

\item{minFinalFeatures}{Integer value for the minimum number of final features selected.}
}
\description{
Adjust variables in a VIVID object
}
