% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid_crit.R
\name{vivid_crit}
\alias{vivid_crit}
\title{Change the information criteria used in VIVID}
\usage{
vivid_crit(vividObj, x, y, metric, lambda = "lambda.1se")
}
\arguments{
\item{vividObj}{An object passed from the vivid() funciton.}

\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{metric}{AIC,  AICc, BIC or EBIC.}

\item{lambda}{Used in cv.glmnet}
}
\description{
Change the information criteria used in VIVID
}
