% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid_plot.R
\name{vivid_plot}
\alias{vivid_plot}
\title{A funciton that creates graphics for the VIVID method}
\usage{
vivid_plot(vividObj, log = TRUE)
}
\arguments{
\item{vividObj}{An object passed from the vivid() funciton.}

\item{log}{A TRUE or FALSE variable that defines whether the values plotted are log-transformed.}
}
\description{
A funciton that creates graphics for the VIVID method
}
