% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid_reg.R
\name{vivid_reg}
\alias{vivid_reg}
\title{Regression method used in the VIVID method}
\usage{
vivid_reg(weight, x, y, crossfold, lambda = "lambda.min")
}
\arguments{
\item{weight}{A p length vector}

\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{crossfold}{Number of crossfold samples.}

\item{lambda}{Used in cv.glmnet.}
}
\description{
Regression method used in the VIVID method
}
