% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivid_split.R
\name{vivid_split}
\alias{vivid_split}
\title{VIVID method when the data must be split into multiple groups.}
\usage{
vivid_split(x, y, bootstraps = 100, cores = parallel::detectCores() -
  1, seed = 1234567, minSize = 2, lambda = "lambda.1se",
  compareMethod = "EBIC", gamma = 1, groups, disjoint = TRUE,
  repFeatures)
}
\arguments{
\item{x}{input matrix, rows represent observations and columns represent
features. column names and row names are recommended.}

\item{y}{response variable, restricted to two classes.}

\item{bootstraps}{integer, the number of bootstraps to run.}

\item{cores}{integer, number of cores to use for parallel processing. VIVID
makes uses of the "future" and "furrr" packages and will set the plan to
"multiprocess". The cores parameter is directly fed to workers arguement of
future::plan.}

\item{seed}{integer, passed to set.seed function, a full L'Ecuyer-CMRG RNG seed (vector of 7 integers).}

\item{minSize}{minimum cluster size to consider when clustering.}

\item{lambda}{Used in cv.glmnet}

\item{compareMethod}{AIC,  AICc, BIC or EBIC.}

\item{gamma}{gamma value for EBIC.}

\item{groups}{Integer value for the total number of groups.}

\item{disjoint}{TRUE or FALSE value on whether the groups are disjoint.}

\item{repFeatures}{Vector of features that are repeated in each group.}
}
\description{
VIVID method when the data must be split into multiple groups.
}
